#14 renamed the folders, also added a home page
