+++ src/components/Pagination/Pagination.js
@@ -3,12 +3,12 @@ import React from "react";
 import { Link, useParams } from "react-router-dom";
 
 const Pagination = ({ maxPage, pageLimit }) => {
-  let { params } = useParams();
-  if (isNaN(params)) params = 1;
-  params = parseInt(params);
+  let { pageNumber } = useParams();
+  if (isNaN(pageNumber)) pageNumber = 1;
+  pageNumber = parseInt(pageNumber);
 
   const getPaginationGroup = () => {
-    let start = Math.floor((params - 1) / pageLimit) * pageLimit;
+    let start = Math.floor((pageNumber - 1) / pageLimit) * pageLimit;
     return new Array(pageLimit).fill().map((_, idx) => start + idx + 1);
   };
 
@@ -23,8 +23,8 @@ const Pagination = ({ maxPage, pageLimit }) => {
         */}
       <div className="pagination">
         {/* previous button */}
-        <Link to={`/page/` + (params - 1)} className={`link-prev ${params === 1 ? "disabled" : ""}`}>
-          <button className={`prev ${params === 1 ? "disabled" : ""}`}>
+        <Link to={`/page/` + (pageNumber - 1)} className={`link-prev ${pageNumber === 1 ? "disabled" : ""}`}>
+          <button className={`prev ${pageNumber === 1 ? "disabled" : ""}`}>
             prev
           </button>
         </Link>
@@ -33,15 +33,15 @@ const Pagination = ({ maxPage, pageLimit }) => {
         {getPaginationGroup().map((item, index) => (
           <button
             key={index}
-            className={`paginationItem ${params === item ? "active" : null}`}
+            className={`paginationItem ${pageNumber === item ? "active" : null}`}
           >
             <span>{item}</span>
           </button>
         ))}
 
         {/* next button */}
-        <Link to={`/page/` + (params + 1)} className={`link-next ${params === maxPage ? "disabled" : ""}`}>
-          <button className={`next ${params === maxPage ? "disabled" : ""}`}>
+        <Link to={`/page/` + (pageNumber + 1)} className={`link-next ${pageNumber === maxPage ? "disabled" : ""}`}>
+          <button className={`next ${pageNumber === maxPage ? "disabled" : ""}`}>
             next
           </button>
         </Link>
