+++ src/pages/home-page/HomePage.css
@@ -1,0 +1,13 @@
+.home-page-main {
+  text-align: center;
+  margin: auto;
+  width: 25%;
+  margin-top: 8em;
+  border: 3px solid green;
+  padding: 10px;
+}
+
+.home-page-center {
+    text-align: center;
+    margin-top: 2em;
+}
